fix(cli): cannot set environment variable `CI=false`
We check the `$CI` environment variable for its presence, not for its
value. This means that setting `CI=false` enables CI mode instead of
disabling it.

This is unexpected. Fix it.

We could have checked for `CI=true` or `CI=1`, but the current behavior
is backwards compatible with crazy things users might already be doing.