fix(cli): cannot set environment variable `CI=false` (#32749)
We check the `$CI` environment variable for its presence, not for its value. This means that setting `CI=false` enables CI mode instead of disabling it.

This is unexpected. Fix it.

We could have checked for `CI=true` or `CI=1`, but the current behavior is backwards compatible with crazy things users might already be doing.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*