fix(integ-runner): `ENOENT` no such file or directory 'recommended-feature-flags.json' (#32750)
In the `integ-runner`, we used to load the recommended feature flags file for `aws-cdk-lib` using a run-time `readFile` based on a `__dirname`.

However, the source files are bundled which makes `__dirname` point to the wrong directory. Instead, load the JSON file directly using a JavaScript `import`, so that the bundler will bundle the JSON file into the source code directly, and we can't fail to load it.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*