chore(cli): remove legacy compat code to support version reporting in v1 apps (#32710)
### Reason for this change

In the first versions of AWS CDK v1 and older, version metadata was injected by the CLI not by the framework.
This changed with the introduction of Cloud Assembly Schema v6.0.0 in AWS CDK v1.65.0 (see [changelog](https://github.com/aws/aws-cdk/releases/tag/v1.65.0)).

The CLI kept compatibility code around to ensure the metadata is injected. This code was introduced via https://github.com/aws/aws-cdk/pull/19010 on Feb 17, 2022.

However this code was intended for v1 and kept around because v1 and v2 were initially developed on the same code base. In the meantime, AWS CDK v1 is EOS and we do not need to support v1 features from v2 anymore.

This change will mean that we will stop reporting metadata for users that use the CLI v2 `>= 2.174.0` (the release removing the compat code) to deploy a Cloud Assembly developed with framework v1 `< 1.65.0`. Otherwise these users are not functionally effected by the change.

We therefore decide to now remove the unused compat code.

### Description of changes

Remove the unused legacy code and adjusted test cases.

### Describe any new or updated permissions being added

n/a

### Description of how you validated changes

Run unit test and CLI integ tests.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*