chore(cli): aliased commands can be converted to cli args (#32806)
This PR does not currently change CLI functionality. The function
`convertToCliArgs` is not used in the CLI yet which is why this PR is
not fixing a regression. It will eventually be used to strongly-type cli
arguments.

Previously, aliased commands, like `cdk ack` instead of `cdk
acknowledge` would fall through the cracks of the generated convert
function. The switch statement was only switching on command names so we
would not store any options associated with an aliased command.
Specifically, `cdk synth --exclusively` would _not_ store the
`exclusively` flag in the ensuing `CliArguments` object because `synth`
is an alias.

Now we do. This is an additional step forward to being able to use
`CliArguments` in `cli.ts`

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*