chore(cli): generate conversion from cdk.json to cli arguments (#32803)
This PR does not change the functionality of the CLI (yet)

It does however articulate a schema for what `cdk.json` should look like
in the future. I'm aware that we honor a slightly different set of rules
in `cdk.json` that _is not documented anywhere_, and we will have to
honor those rules ad-hoc. However, this will hopefully move us towards a
strongly-typed future where `cdk.json` contents mirror CLI argument
options.

- global options are specified at the base level of `cdk.json`
- command specific options will be prefixed by their command name. NOTE:
some options are honored at the base level today. I will have to, in a
separate PR, find each of these instances and take care of them but
ensuring we still map them to the correct place in `CliArguments`.

```json
{
  "app": "npx ts-node -P tsconfig.json --prefer-ts-exts src/main.ts",
  "output": "cdk.out",
  "build": "npx projen bundle",
  "watch": {
      "exclude": [
        "README.md",
        "cdk*.json",
        "**/*.d.ts",
        "**/*.js",
        "tsconfig.json",
        "package*.json",
        "yarn.lock",
        "node_modules"
    ]
}
```

This will turn into the following `CliArgument` object:

```ts
{
  globalOptions: {
    app:  'npx ts-node -P tsconfig.json --prefer-ts-exts src/main.ts',
    output: 'cdk.out',
    build: 'npx projen bundle',
    watch: {
      exclude: [
          "README.md",
          "cdk*.json",
          "**/*.d.ts",
          "**/*.js",
          "tsconfig.json",
          "package*.json",
          "yarn.lock",
          "node_modules",
      ],
   },
};
```

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*

---------

Co-authored-by: Momo Kornher <kornherm@amazon.co.uk>