chore(cli): synth is now default command name, synthesize is alias (#32823)
As a product we have standardized on `cdk synth`. This change reflects that while not changing behavior of the CDK CLI at all. It will however matter for the behavior of a future feature where we allow defaults specified in a schematic way in `cdk.json`. The payoff is that instead of requiring `synthesize: { ... }` we instead will require `synth: { ... }`.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*