chore(cli): rename `_` to `command` in `UserInput` (#32822)
This does not change CLI functionality because `UserInput` is not in use yet. Since we have full control over `UserInput`, i.e. we control the input functions from the CLI or `cdk.json`, we can rename properties however we like. `_` is a relic of `yargs`, we do not need to maintain that convention.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*