feat(ecs): enable Enhanced Observability for Container Insights (#32622)
### Issue # (if applicable)

Closes #32618.

### Reason for this change


Container Insights with enhanced observability was recently announced and released for ECS. Cloudformation and the L1 CFN construct support setting this at cluster level. This adds the Container Insights with enhanced observability as an option for the existing L2 Cluster construct. 

### Description of changes


ENHANCED was added to the ContainerInsights enum. A new containerInsightsV2 property was added to the Cluster construct with the type ContainerInsights. The existing containerInsights property was marked as deprecated.


### Description of how you validated changes


Unit and Integration tests were added.

### Checklist
- [X] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*