chore(prlint): fail prlinter on codecov failures, with exemption label (#32674)
Our current process for codecov statuses is that they block the merge of a PR if the are failing. This PR attempts to move this failure further forward in the PR process, so that the PRlinter fails and alerts the author. It also means that we can include an exemption label to merge a PR that we allow to be in breach of the codecov report. Currently, to do so we must change the branch protection rules.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*