chore: disable silent tests
Remove `silent: true` from the test config. `console.log()` is a
time-honored debugging tradition, and the `silent: true` config makes
that not work in a non-obvious way.

A nice behavior that many other testing frameworks have is to only
print the console output of tests that fail. Use the "Jest Environments"
feature to build an environment that replaces `console.log()` etc and
buffers their output to only print it for failing tests.