chore: don't add the same Aspect more than once (#32880)
There is no observable behavior from adding the same Aspect more than
once at the same priority, but we are adding multiple
`AspectApplication`s for it (and then only executing once).

Make the proptest results a bit easier to read by not allowing this in
the first place.

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*