chore: re-enable console output in tests (#32883)
Remove `silent: true` from the test config. `console.log()` is a time-honored debugging tradition, and the `silent: true` config makes that not work in a non-obvious way.

A nice behavior that many other testing frameworks have is to only print the console output of tests that fail. Use the "Jest Environments" feature to build an environment that replaces `console.log()` etc and buffers their output to only print it for failing tests.

This achieves the same goals of not polluting the terminal too much while running tests, while still allowing the ability to see the output of `console.log`.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*