chore: stop proptest failures by giving fewer guarantees (#32885)
The previous Aspect proptests were trying to assert ordering guarantees that the code didn't actually guarantee. This led to sporadic failures as the tests found scenarios in which the assertions were broken.

In this PR, cut down on the guarantees that we make: only guarantee those things on aspect applications that exist at the start of iteration, and not on aspect applications that are added during iteration.

This is not ideal, but it accurately reflects the current code and it stops intermittent failures. We will revisit tighter guarantees on aspects that are added during iteration in a later PR.

Depends-On: https://github.com/aws/aws-cdk/pull/32884

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*