chore(cli): also suppress test logging for CLI
The CLI tests are also quite noisy. Suppress logging for those as well
by using the `jest-bufferedconsole` environment.

Enhance the buffered-console environments by also capturing writes to
`process.stdout` and `process.stderr`.

Yes, the files are copy/pasted between the CLI and the library. I don't
see an easy way around this, and the code is going to be lifted
into a separate repo soon anyway.