chore(cli): also suppress test logging for CLI (#32952)
The CLI tests are also quite noisy. Suppress logging for those as well by using the `jest-bufferedconsole` environment.

Enhance the buffered-console environments by also capturing writes to `process.stdout` and `process.stderr`.

Yes, the files are copy/pasted between the CLI and the library. I don't see an easy way around this, and the code is going to be lifted into a separate repo soon anyway.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*