chore(toolkit): programmatic toolkit for the AWS CDK initial code (#32919)
### Description of changes

Initial code for the Programmatic Toolkit. This won't be released just yet.
Contains a mix of extensions and hard copies to the current CLI code.
After this PR we are moving the appropriate tests over from the CLI.

### Describe any new or updated permissions being added

n/a

### Description of how you validated changes

For the changes to `aws-cdk` we run the existing tests and the integration tests.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*