chore: ensure docker file runs as non root user (#32976)
### Reason for this change

Fix Code Scanner issue

```
By not specifying a USER, a program in the container may run as 'root'. This is a security hazard.
If an attacker can control a process running as root, they may have control over the container.
Ensure that the last USER in a Dockerfile is a USER other than 'root'.
```

### Description of changes

Create a new group and attach the user to the group. The dockerfile already gives necessary permissions with statements like `chmod 777`

### Description of how you validated changes

N/A

### Checklist
- [ ] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*

(cherry picked from commit ddaad470c5da4ea85789dc53a27af58832ae0e1c)