feat(cloudfront-origins): list access level for 404 response (#32059)
### Issue # (if applicable)

Closes #13983.
Closes #31689.

### Reason for this change

When we want to receive HTTP 404 response where the requested object does not exist,
s3:ListBucket permission is needed in the S3 bucket policy.

Unlike `errorResponses` to convert 403 response to 404, This is useful to distinguish between responses blocked by WAF (403) and responses where the file does not exist (404).

### Description of changes

Added a new `AccessLevel.LIST` to allow s3:ListBucket.

### Description of how you validated changes

Unit test and integration test. The integ test also tests the response is 404.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*