docs(appsync): correct the schema definition for the sample code  (#32889)
### Issue #32848

Closes #32848

Reason for this change
The current sample schema is incorrect and causes the stack deployment to fail.

Description of changes
I modified the sample GraphQL schema so that it is successfully deployed.

Describe any new or updated permissions being added
<!— What new or updated IAM permissions are needed to support the changes being introduced ? -->

Description of how you validated changes
I was able to successfully deploy the stack after making the changes I already proposed in the PR.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*