docs(contribution-guidelines): add guidance on a common "error" (#32983)
### Issue # (if applicable)

Closes #<issue number here>.

### Reason for this change

Anecdotally, contributors often encounter the "This branch is out-of-date with the base branch" message, which can be confusing. Since I couldn’t find a clear explanation, I sought clarification from one of the admins in [this comment](https://github.com/aws/aws-cdk/pull/32889#issuecomment-2597545560). I’ve summarized their guidance to help other contributors navigate this issue more easily.




### Description of changes

Added clarification on a common "error" in the contributor guidelines.




### Describe any new or updated permissions being added




### Description of how you validated changes

An admin provided guidance on the issue, and it resolved the problem effectively in my case.




### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*