chore: stop failing the PR linter on changes requested (#32987)
Almost every PR immediately looks like it's failing with a red cross, because the PR linter fails if it is requesting changes.

The "Changes Requested" review by itself is enough to prevent a PR from getting merged by the Mergify config, so we don't actually need to fail the PR linter as well.

Instead: the PR linter succeeds if it runs to the end, and it may request changes on the PR. If it fails, then it's because it was unable to do its job for some reason (that should and will still block merging, so we are not accidentally failing open if something is wrong with the linter).

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*