chore: fix memory overflow for buffered console (#32986)
Since the buffered console captures stdout/stderr, in some call
sequences it keeps recursing forever and overflows memory.

It does not repro in this repository, but it repros in a different one.

The fix is to stop capturing while we print results.

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*

(cherry picked from commit 776620d042c25f34cc2315d8d48917d60aa8bbf2)