docs(contributing): add introduction to contributing and update flowchart to include major and minor contributions (#32619)
### Issue # (if applicable)

Closes #<issue number here>.

### Reason for this change

Make it easier for those new to contributing and for those submitting small contributions to get started.

### Description of changes

Provide introductory information on what a contribution is. Add a new definition of *major* and *minor* contributions. The purpose of these new terms are to distinguish between two different types of contributions for two reasons: (1) Provide a clearer and simpler path to contributing for those submitting small changes like doc improvements or bug fixes; and (2) Provide a separate path of submitting an RFC to discuss implementation details before someone puts in the time and effort of submitting a major contribution.

I also updated the flowchart to highlight these two paths and show how they differ. 

Next steps after this revision are to improve the "getting started" documentation and restructure content based on the updated flowchart. 

### Describe any new or updated permissions being added

<!— What new or updated IAM permissions are needed to support the changes being introduced ? -->

### Description of how you validated changes



### Checklist
- [ ] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*