chore: make tests work with new SDKv3 version
In a recent upgrade of the SDKv3, the INI credential provider has gotten
a private copy of the STS Client, to make AssumeRole calls with. This
means the STS Client used by the INI provider is now unmockable.

In #31702, we migrated the tests of the credentials chain mostly to
using mocks, as opposed to what it was doing in the past: intercepting
network traffic and emulating a fake STS Endpoint. These tests now start
failing due to an SDK upgrade.

Fortunately, most of the old code was still there, so in this change
I'm ripping out the STS Mocking and doing a couple minor changes;
the tests now pass again.