fix(cli): disallow import of internal cli libraries
Previously it was possible to import any subpaths from the `aws-cdk` package ("the CLI"). This was never intended to be allowed, or supported. In practice, these subpaths imports allowed users to depend on internal CLI APIs that are not intended for public usage and do not receive the same backwards compatibility guarantees as other parts of the AWS CDK.

**With this change we are explicitly disallowing unsanctioned subpath imports.**

We are currently in the process of making most CLI features available through a new Programmatic Toolkit library. Please see the [respective RFC](https://github.com/aws/aws-cdk-rfcs/pull/654) and let us know if you have a use case that is not currently covered by the proposed feature set.

In order to not immediately break all customers using unsanctioned subpath imports, we have identified a subset of symbols that we will keep exporting in the short-time future. **You are still very strongly encouraged to move off any of these features asap.** We are actively considering to emit warnings and enact brown-outs to inform users of this removal.