chore: make tests work with new SDKv3 version (#33016)
In a recent upgrade of the SDKv3, the INI credential provider has gotten
a private copy of the STS Client, to make AssumeRole calls with. This
means the STS Client used by the INI provider is now unmockable.

In #31702, we migrated the tests of the credentials chain mostly to
using mocks, as opposed to what it was doing in the past: intercepting
network traffic and emulating a fake STS Endpoint. These tests now start
failing due to an SDK upgrade.

Fortunately, most of the old code was still there, so in this change I'm
ripping out the STS Mocking and doing a couple minor changes; the tests
now pass again.

This PR also upgrades the SDKv3 version at the same time, some other
packages that needed to be upgraded along with this as well
(`@smithy/middleware-endpoint` and `cdk-assets` which covers a new enum
value for the S3 client).

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>
(cherry picked from commit d4845ce0b32913dbba465efe47d59411809b0274)