chore: add integ tests that validate full proxy traversal
Add a CLI integ test to validate that `cdk deploy` works in a fully
network-isolated environment, with only a proxy to go through.

This validates that no parts of the CLI setup ignore the proxy
configuration, which would otherwise be hard to test.

We achieve the network isolation by running the code inside a Docker
container where we use `iptables` to drop all network traffic that
doesn't go through the Docker host, where we run `mitmproxy`.