chore(prlint): all prefixes must be lowercase + better failure message (#33102)
Updates the PR linter to remove `Revert` as a valid title prefix, in favor of `revert`. To be fair, this is a stylistic change, but I think its unclean to have every other valid prefix as lowercase except for `Revert`. The reason is because github automatically titles revert PRs with `Revert`, but I don't think it's much effort for engineers to update the title so it fits in the style with everything else.

In addition, I've updated the failure message to make it clear what it's looking for, rather than a link to conventionalcommits.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*

(cherry picked from commit 809a7f0aa27cc0491a623ee80d9447291c83f3a4)